Test implementation of FClause based on Multiset.
In master, FClause is implemented using Lists, but equality is defined
as multiset equality. For this reason, the current hashCode
implementation of FClause is very basic and has a lot of collisions in
practice (clauses of equal length have the same hashCode).

This patch extends the Multiset data structure in GAPT and uses it in
the implementation of FClause. This should greatly improve the
performance of all algorithms using clause sets.